requirements.txt
python-binance
streamlit
pandas
numpy
plotly
python-dotenv